refactor(shared-index): extract model catalog constants
shared-index.ts is the global SharedIndex DO. Constants like
FALLBACK_MODELS, WORKERS_AI_MODELS, FLUX_IMAGE_MODEL, etc. were
exported from it purely so coding-agent.ts could import them — wrong
shape (coding-agent shouldn't depend on a DO file for static data).
Extract to src/model-catalog.ts. Per the improve-codebase-architecture
audit.
